#120 Refactored integration tests
