+++ tests/ci_tests/test_training_autopilot.py
@@ -3,15 +3,13 @@ from datetime import datetime
 import pytest
 import exasol.bucketfs as bfs
 
-from tests.ci_tests.fixtures.prepare_environment_fixture import CITestEnvironment
+from tests.fixtures.prepare_environment_fixture import CITestEnvironment
 from tests.ci_tests.utils.autopilot_training import AutopilotTestTraining
-from tests.ci_tests.utils.checkers import skip_if_aws_credentials_not_set
 from tests.ci_tests.utils.parameters import reg_model_setup_params, \
     cls_model_setup_params
 from tests.ci_tests.utils.queries import DatabaseQueries
 
 
-@skip_if_aws_credentials_not_set
 @pytest.mark.parametrize("db_conn", [bfs.path.StorageBackend.onprem, bfs.path.StorageBackend.saas], indirect=True)
 def test_train_autopilot_regression_job(db_conn, setup_ci_test_environment):
     curr_datetime = datetime.now().strftime("%y%m%d%H%M%S")
@@ -26,7 +24,6 @@ def test_train_autopilot_regression_job(db_conn, setup_ci_test_environment):
         job_name, reg_model_setup_params, setup_ci_test_environment)
 
 
-@skip_if_aws_credentials_not_set
 @pytest.mark.parametrize("db_conn", [bfs.path.StorageBackend.onprem, bfs.path.StorageBackend.saas], indirect=True)
 def test_train_autopilot_classification_job(db_conn, setup_ci_test_environment):
     curr_datetime = datetime.now().strftime("%y%m%d%H%M%S")
