+++ tests/ci_tests/utils/autopilot_polling.py
@@ -1,6 +1,6 @@
 import os
 
-from tests.ci_tests.fixtures.prepare_environment_fixture import CITestEnvironment
+from tests.fixtures.prepare_environment_fixture import CITestEnvironment
 
 
 class AutopilotTestPolling:
