+++ tests/ci_tests/utils/autopilot_prediction.py
@@ -1,4 +1,4 @@
-from tests.ci_tests.fixtures.prepare_environment_fixture import CITestEnvironment
+from tests.fixtures.prepare_environment_fixture import CITestEnvironment
 
 
 class AutopilotTestPrediction:
