+++ tests/conftest.py
@@ -1,12 +1,8 @@
 pytest_plugins = [
-    "tests.unit_tests.fixtures.aws_connection_fixture",
-    "tests.integration_tests.fixtures.database_connection_fixture",
-    "tests.integration_tests.fixtures.build_language_container_fixture",
-    "tests.integration_tests.fixtures.setup_database_fixture",
-    "tests.integration_tests.fixtures.deploy_scripts_fixture",
-
-    "tests.ci_tests.fixtures.database_connection_fixture",
-    "tests.ci_tests.fixtures.build_language_container_fixture",
-    "tests.ci_tests.fixtures.prepare_environment_fixture",
-    "tests.ci_tests.fixtures.setup_ci_test_environment"
+    "tests.fixtures.aws_connection_fixture",
+    "tests.fixtures.database_connection_fixture",
+    "tests.fixtures.build_language_container_fixture",
+    "tests.fixtures.setup_database_fixture",
+    "tests.prepare_environment_fixture",
+    "tests.setup_ci_test_environment"
 ]
